cruntime: Fix (List|Stop|Kill)Containers behavior when no containers match

+++ pkg/minikube/cruntime/docker.go
@@ -89,16 +89,30 @@ func (r *Docker) ListContainers(filter string) ([]string, error) {
 	if err != nil {
 		return nil, err
 	}
-	return strings.Split(content, "\n"), nil
+	var ids []string
+	for _, line := range strings.Split(content, "\n") {
+		if line != "" {
+			ids = append(ids, line)
+		}
+	}
+	return ids, nil
 }
 
-// KillContainers forcibly removes a running pod based on ID
+// KillContainers forcibly removes a running container based on ID
 func (r *Docker) KillContainers(ids []string) error {
+	if len(ids) == 0 {
+		return nil
+	}
+	glog.Infof("Killing containers: %s", ids)
 	return r.Runner.Run(fmt.Sprintf("docker rm -f %s", strings.Join(ids, " ")))
 }
 
-// StopContainers stops a running pod based on ID
+// StopContainers stops a running container based on ID
 func (r *Docker) StopContainers(ids []string) error {
+	if len(ids) == 0 {
+		return nil
+	}
+	glog.Infof("Stopping containers: %s", ids)
 	return r.Runner.Run(fmt.Sprintf("docker stop %s", strings.Join(ids, " ")))
 }
